requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
beautifulsoup4==4.12.3
requests>=2.32.3
python-dotenv==1.0.1
pydantic==2.6.1
aiohttp==3.9.3
cachetools==5.3.2
sqlalchemy==2.0.25
# Database driver
psycopg2-binary==2.9.9  # PostgreSQL
transformers==4.37.2
accelerate==0.27.2
smolagents[openai]==1.13.0
markdownify>=0.14.1
playwright==1.52.0
# PyTorch needs to be installed separately using:
# For CPU-only:
# pip install torch torchvision torchaudio
# For CUDA 12.1:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

langchain==0.1.4
langchain-community>=0.0.14,<0.1
langchain-core==0.1.17
nest_asyncio 